remove more deprecated overrides

+++ Classes/EditGroupViewController.m
@@ -172,7 +172,7 @@ static PassDropAppDelegate *app;
     return kdbGroup;
 }
 
-- (void)iconSelected:(UIImage*)icon withId:(NSInteger)iconId {
+- (void)iconSelected:(UIImage*)icon withId:(uint32_t)iconId {
     self.neuIcon = icon;
     self.newIconId = iconId;
     [[self tableView] reloadData];
